refactor: 회원 탈퇴 AuthController 에서 수행하도록 변경 - #170
